factor out evaluating koka out of the compilation function
Previously running Koka code with the -e flag was done at the end of
mainMode used for compiling, but that made it run inside the color 
printer, which made it emit a reset ANSI sequence after evaluating Koka. 
I moved the evaluation step out of mainMode into runWithLSArgs so that
it's done after the color printer is "closed".